[REFACTOR] Rename UserSummary to UsersSummary
